refactor : 템플릿 앞면(텍스트입력) 컴포넌트 별도 분리
- 추후, 초대장 이미지화를 위해 이미지화 될 (텍스트가 입력된)템플릿 컴포넌트를 별도 분리하였습니다!

ref : #11
